Create new dir/namespace `model`, migrate Item and its descendants there, make the classes importable directly from the namespace
