+++ phibes/model/schema.py
@@ -1,7 +1,7 @@
 """
-Implementation of Template
+Implementation of Schema
 
-A Template is an entry in a Locker.
+A Schema is used to enforce format on other entities, like Tags.
 """
 
 # Built-in library packages
@@ -9,11 +9,11 @@ A Template is an entry in a Locker.
 # Third party packages
 
 # In-project modules
-from . crypto import CryptImpl
-from . item import Item
+from phibes.lib.crypto import CryptImpl
+from phibes.model import Item
 
 
-class Template(Item):
+class Schema(Item):
 
     def __init__(self, crypt_obj: CryptImpl, name: str, content: str = ""):
         super().__init__(crypt_obj, name, content)
