+++ phibes/model/secret.py
@@ -1,7 +1,7 @@
 """
-Implementation of Schema
+Implementation of Secrets
 
-A Schema is used to enforce format on other entities, like Tags.
+A Secret is an entry in a Locker.
 """
 
 # Built-in library packages
@@ -9,11 +9,11 @@ A Schema is used to enforce format on other entities, like Tags.
 # Third party packages
 
 # In-project modules
-from . crypto import CryptImpl
-from . item import Item
+from phibes.lib.crypto import CryptImpl
+from phibes.model import Item
 
 
-class Schema(Item):
+class Secret(Item):
 
     def __init__(self, crypt_obj: CryptImpl, name: str, content: str = ""):
         super().__init__(crypt_obj, name, content)
